added sharedComponents to homeView
