+++ app/src/main/java/com/example/clicker/presentation/home/views/HomeComponents.kt
@@ -137,9 +137,6 @@ import com.example.clicker.util.Response
                     showLogoutDialog = {
                         showLogoutDialog()
                     },
-                    login = {
-                        loginWithTwitch()
-                    },
                     userIsLoggedIn =userIsAuthenticated,
 
                     screenDensity =screenDensity,
